kjhc

+++ Science.Api/Science.Service/Services/StudentService.cs
@@ -1,4 +1,5 @@
 ﻿using Science.Data.IRepositories;
+using Science.DTO.Student;
 using Science.Entity;
 using Science.Service.IServices;
 
@@ -13,14 +14,20 @@ namespace Science.Service.Services
             this.studentRepository = studentRepository;
         }
 
-        public async Task Create(Student student)
+        public async Task CreateAsync(Student student)
         {
             await studentRepository.Create(student);
         }
 
-        public async Task Delete(Student student)
+        public async Task<bool> Delete(Guid Id)
         {
-            await studentRepository.Delete(student);
+            var student = await studentRepository.GetById(Id);
+
+            if(student == null) { return false; }
+
+            studentRepository.Delete(student);
+
+            return true;
         }
 
         public async Task<IEnumerable<Student>> GetAllAsync()
@@ -28,9 +35,14 @@ namespace Science.Service.Services
            return await studentRepository.GetAllAsync();
         }
 
-        public async Task<Student> GetByServiesId(Guid id)
+        public async Task<Student> GetByIdAsync(Guid id)
         {
             return await studentRepository.GetById(id);
+        }
+
+        public async Task SaveChangesAsync()
+        {
+            await studentRepository.SaveChangesAsync();
         }
 
         public void Update(Student student)
